Extract private classes (#40)
* extract observinglabel

* extract score class
